requirements.txt
# Core ML/DL
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
numpy>=1.21.0
scipy>=1.7.0

# Data processing
pandas>=1.3.0
Pillow>=8.0.0
opencv-python>=4.5.0

# Training utilities
tqdm>=4.60.0
tensorboard>=2.8.0

# Evaluation and metrics
editdistance>=0.6.0
jiwer>=2.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Dashboard (Phase 3)
streamlit>=1.30.0

# CTC decoding (optional, for beam search with LM)
# ctcdecode>=1.0.0  # Requires manual installation from source

# Configuration
pyyaml>=6.0
omegaconf>=2.1.0

# Logging
wandb>=0.12.0

# MediaPipe for feature extraction (optional)
mediapipe>=0.9.0

# Language modeling (optional)
kenlm>=0.0.0  # Requires manual installation

